part2/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

pydantic==2.5.0
pydantic-settings==2.1.0

google-generativeai==0.3.2

#issues finding the compartible version: will let pip figure it out 
#figured out which versions were suitable
# langchain
# langchain-community
langchain-google-genai
#langgraph


asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic  #issues
sqlalchemy[asyncio]==2.0.25

pandas==2.1.3
openpyxl==3.1.2
xlrd==2.0.1

httpx==0.25.2
requests==2.31.0

asyncio==3.4.3

structlog==23.2.0

pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

python-dotenv==1.0.0

marshmallow==3.20.2
marshmallow-sqlalchemy==0.29.0

# Caching (for performance)
redis==5.0.1

# Background tasks
celery==5.3.4


XlsxWriter==3.2.0